+++ examples/platformer_2d.rs
@@ -236,6 +236,8 @@ fn setup_player(mut commands: Commands) {
             tilt_offset_angvel: 5.0,
             tilt_offset_angacl: 500.0,
             turning_angvel: 10.0,
+            height_change_impulse_for_duration: 0.02,
+            height_change_impulse_limit: 40.0,
         },
         ..Default::default()
     });
@@ -328,11 +330,16 @@ fn apply_controls(
         .into_iter()
         .any(|key_code| keyboard.pressed(key_code));
 
+    let crouch = [KeyCode::Down, KeyCode::LControl, KeyCode::RControl]
+        .into_iter()
+        .any(|key_code| keyboard.pressed(key_code));
+
     for mut controls in query.iter_mut() {
         *controls = TnuaPlatformerControls {
             desired_velocity: if turn_in_place { Vec3::ZERO } else { direction },
             desired_forward: direction.normalize_or_zero(),
             jump: jump.then(|| 1.0),
+            float_height_offset: if crouch { -0.9 } else { 0.0 },
         };
     }
 }