+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `float_height_offset` control for crouching. Also add:
+  - `height_change_impulse_for_duration` and `height_change_impulse_limit`
+    settings for controling a boost that would be added for crouching and
+    getting back up.
+  - `standing_offset` field to `TnuaPlatformerAnimatingOutput` to assist in
+    applying crouching/crawling animation.
 
 ## 0.3.0 - 2023-05-13
 ### Changed